+++ src/components/About/About.js
@@ -1,0 +1,10 @@
+function About() {
+  return (
+    <div>
+      <h1>About</h1>
+      <p>This is About page</p>
+    </div>
+  );
+}
+
+export default About;
