+++ src/components/Header/Header.css
@@ -2,3 +2,24 @@
   color: gray;
   width: 50%;
 }
+
+.header > ul {
+  display: flex;
+}
+
+a:visited {
+  color: black;
+}
+
+a:hover {
+  color: blueviolet;
+}
+
+.header > ul > li > a {
+  text-decoration: none;
+}
+
+.header > ul > li {
+  font-weight: 700;
+  margin: 1em;
+}
