+++ src/components/Header/Header.js
@@ -1,10 +1,15 @@
 import './Header.css';
+// eslint-disable-next-line
+import { Link } from 'react-router-dom';
 
 function Header() {
   return (
     <div className="header">
-      <h1>Todos</h1>
-      <p>Your todos are listed below</p>
+      <ul>
+        <li><Link to="/">Home</Link></li>
+        <li><Link to="/about">About</Link></li>
+        <li><Link to="/profile">Profile</Link></li>
+      </ul>
     </div>
   );
 }
