+++ src/components/Profile/Profile.js
@@ -1,0 +1,10 @@
+function Profile() {
+  return (
+    <div>
+      <h1>Profile</h1>
+      <p>This is profile page</p>
+    </div>
+  );
+}
+
+export default Profile;
